refactor scoring module into post module
