+++ app/modules/posts/scoring/global-brain-service.ts
