+++ app/modules/posts/scoring/scoring-utils.ts
