+++ app/modules/posts/scoring/vote-service.ts
@@ -1,8 +1,8 @@
 import { type Transaction } from 'kysely'
-import { sendVoteEvent } from '#app/modules/scoring/global-brain-service.ts'
 import { Direction, type VoteState } from '#app/types/api-types.ts'
 import { type DBVoteEvent } from '#app/types/db-types.ts'
 import { type DB } from '#app/types/kysely-types.ts'
+import { sendVoteEvent } from './global-brain-service.ts'
 import { insertVoteEvent } from './vote-repository.ts'
 
 export function defaultVoteState(postId: number): VoteState {
